relationships/requirements.txt
bottle==0.12.7
cloudify-plugins-common==3.4
cloudify-rest-client==3.4
Jinja2==2.7.2
MarkupSafe==0.23
networkx==1.8.1
pika==0.9.14
proxy-tools==0.1.0
python-consul==0.6.1
requests==2.7.0
requests-toolbelt==0.7.0
six==1.10.0
